converter moved to sql.go
