+++ typeutils/flatten.go
@@ -54,14 +54,8 @@ func (f *FlattenerImpl) flatten(key string, value any, destination types.Record)
 			return fmt.Errorf("error marshaling array with key[%s] and value %v: %v", key, value, err)
 		}
 		destination[key] = string(b)
-	case reflect.Bool, reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64,
-		reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64,
-		reflect.Float32, reflect.Float64, reflect.String:
-		destination[key] = value
 	default:
-		if !f.omitNilValues || value != nil {
-			destination[key] = fmt.Sprint(value)
-		}
+		destination[key] = value
 	}
 
 	return nil
